Changed some things
The multidimensional vector of possibilities was changed to a packed array to avoid confusion when referencing. Made the output assigned through a generating always blocks. You know it is done when the done signal goes high. I am having trouble getting it to switch states after it counts 81 final cells