Expanded for loops and added automatic
I expanded the for loops so that they are not for loops. I learned about automatic tasks and functions so I implemented that into the code. This update added about 500 lines, but reduced LUT count by a decent amount. I still need to do some major reconstruction to get the LUT count under 110,000.