Update newSud.v
Fixed some major logic errors (there are still some more tho) and added a new output called borked to go high when the puzzle is unsolvable. The main thing I added was the reg called backtrack so the state machine can know which way it was transversing the cells during the previous clock cycle. I am currently doing simulations to get this module working right. I still need to create a squarechecker function. BUT, if this module works well, I will merge it into the main branch because it actually can fit on to the Nexys Video. Also I need to learn how to use branches and how to merge stuff on Github